requirements.txt
# Core ML/AI Framework
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.35.0
accelerate>=0.24.0
datasets>=2.14.0

# Distributed Training
deepspeed>=0.12.0
peft>=0.7.0  # LoRA/QLoRA
bitsandbytes>=0.41.0

# Multimodal & Vision
timm>=0.9.12
opencv-python>=4.8.1
pillow>=10.0.0
decord>=0.6.0
av>=10.0.0

# Audio Processing
librosa>=0.10.1
soundfile>=0.12.1
whisper>=1.1.10
demucs>=4.0.1

# Video Processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3

# Object Detection & Segmentation
ultralytics>=8.0.0  # YOLO
segment-anything>=1.0
detectron2>=0.6

# Experiment Tracking
wandb>=0.15.12
mlflow>=2.7.1
tensorboard>=2.14.1

# Utilities
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.1.0
matplotlib>=3.7.0
seaborn>=0.12.0
tqdm>=4.66.0
omegaconf>=2.3.0
hydra-core>=1.3.0
rich>=13.6.0
typer>=0.9.0

# Development
pytest>=7.4.0
black>=23.9.0
flake8>=6.1.0
pre-commit>=3.4.0

# Optional: Experimental
diffusers>=0.21.4
controlnet-aux>=0.0.6
xformers>=0.0.22  # Memory efficient attention
